Adding WebApiClient to FirstApi Repository

+++ FirstWebAPI/Models/RepositoryEmployee.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.OpenApi.Extensions;
 
 namespace FirstWebAPI.Models
 {
@@ -18,28 +19,93 @@ namespace FirstWebAPI.Models
             Employee employee=_context.Employees.Find(id);
             return employee;
         }
-        public Employee UpdateEmployee(Employee updatedEmployee) 
+        public int UpdateEmployee(Employee updatedEmployee) 
         {
+            EntityState es = _context.Entry(updatedEmployee).State;
+            Console.WriteLine($"EntityState b4 Update :{es.GetDisplayName()}");
+
             _context.Employees.Update(updatedEmployee);
-           // Console.WriteLine(_context.Entry(updatedEmployee).State); 
+
+            es = _context.Entry(updatedEmployee).State;
+            Console.WriteLine($"EntityState After Update :{es.GetDisplayName()}");
+
             _context.SaveChanges();
-            return updatedEmployee;
+
+            es = _context.Entry(updatedEmployee).State;
+            Console.WriteLine($"EntityState After SaveChanges :{es.GetDisplayName()}");
+
+            return 1;
         }
         public void AddNewEmployee(Employee employee)
         {
-            _context.Employees.Add(employee);
-            _context.SaveChanges();
+            Employee? foundEmp = _context.Employees.Find(employee.EmployeeId);
+            if(foundEmp != null)
+            {
+                throw new Exception("Failed to add new employee.Duplicate Id");
+            }
+            EntityState es = _context.Entry(employee).State;
+            Console.WriteLine($"EntityState b4 Add :{es.GetDisplayName()}");
+
+            _context.Employees.Add(employee); 
+
+            es = _context.Entry(employee).State;
+            Console.WriteLine($"EntityState b4 Add :{es.GetDisplayName()}");
+
+            _context.SaveChanges();   
+
+            es = _context.Entry(employee).State;
+            Console.WriteLine($"EntityState b4 Add :{es.GetDisplayName()}");
         }
         public int DeleteEmployee(int id)
         {
 
-            var employee = _context.Employees.Find(id);
+            Employee? employee = _context.Employees.Find(id);
+            EntityState es = EntityState.Detached;
+            int result = 0;
+            if (employee != null)
+            {
+                es = _context.Entry(employee).State;
+                Console.WriteLine($"EntityState Before Delete :{es.GetDisplayName()}");
 
-            _context.Employees.Remove(employee);
-            _context.SaveChanges();
+                _context.Employees.Remove(employee);
 
-            return 1;
+                es = _context.Entry(employee).State;
+                Console.WriteLine($"EntityState after Delete :{es.GetDisplayName()}");
 
+                result = _context.SaveChanges();
+
+                es = _context.Entry(employee).State;
+                Console.WriteLine($"EntityState after SaveChanges :{es.GetDisplayName()}");
+            }
+            return result;
+
+        }
+        public EmployeeViewModel MapEmployeeToEmpViewModel(Employee employee)
+        {
+            return new EmployeeViewModel
+            {
+                EmpId = employee.EmployeeId,
+                FirstName = employee.FirstName,
+                LastName = employee.LastName,
+                Title = employee.Title,
+                BirthDate = employee.BirthDate,
+                HireDate = employee.HireDate,
+                City = employee.City,
+                ReportsTo = employee.ReportsTo > 0 ? employee.ReportsTo : null
+            };
+        }
+        public Employee MapEmpViewModelToEmployee(EmployeeViewModel empViewModel)
+        {
+            return new Employee
+            {
+                FirstName = empViewModel.FirstName,
+                LastName = empViewModel.LastName,
+                Title = empViewModel.Title,
+                BirthDate = empViewModel.BirthDate,
+                HireDate = empViewModel.HireDate,
+                City = empViewModel.City,
+                ReportsTo = empViewModel.ReportsTo > 0 ? empViewModel.ReportsTo : null
+            };
         }
     }
 }
